Remove zarr.py
