+++ virtualizarr/writers/kerchunk.py
@@ -1,15 +1,17 @@
 import base64
 import json
-from typing import cast
+from typing import Any, cast
 
 import numpy as np
 from xarray import Dataset
 from xarray.coding.times import CFDatetimeCoder
 from xarray.core.variable import Variable
+from zarr.abc.codec import ArrayArrayCodec, BytesBytesCodec
+from zarr.core.metadata import ArrayV2Metadata, ArrayV3Metadata
 
+from virtualizarr.codecs import _get_codec_config, extract_codecs
 from virtualizarr.manifests.manifest import join
 from virtualizarr.types.kerchunk import KerchunkArrRefs, KerchunkStoreRefs
-from virtualizarr.zarr import convert_v3_to_v2_metadata, to_kerchunk_json
 
 
 class NumpyEncoder(json.JSONEncoder):
@@ -88,6 +90,57 @@ def remove_file_uri_prefix(path: str):
         return path
 
 
+def convert_v3_to_v2_metadata(
+    v3_metadata: ArrayV3Metadata, fill_value: Any = None
+) -> ArrayV2Metadata:
+    """
+    Convert ArrayV3Metadata to ArrayV2Metadata.
+
+    Parameters
+    ----------
+    v3_metadata : ArrayV3Metadata
+        The metadata object in v3 format.
+    fill_value : Any, optional
+        Override the fill value from v3 metadata.
+
+    Returns
+    -------
+    ArrayV2Metadata
+        The metadata object in v2 format.
+    """
+    import warnings
+
+    array_filters: tuple[ArrayArrayCodec, ...]
+    bytes_compressors: tuple[BytesBytesCodec, ...]
+    array_filters, _, bytes_compressors = extract_codecs(v3_metadata.codecs)
+
+    # Handle compressor configuration
+    compressor_config: dict[str, Any] | None = None
+    if bytes_compressors:
+        if len(bytes_compressors) > 1:
+            warnings.warn(
+                "Multiple compressors found in v3 metadata. Using the first compressor, "
+                "others will be ignored. This may affect data compatibility.",
+                UserWarning,
+            )
+        compressor_config = _get_codec_config(bytes_compressors[0])
+
+    # Handle filter configurations
+    filter_configs = [_get_codec_config(filter_) for filter_ in array_filters]
+    v2_metadata = ArrayV2Metadata(
+        shape=v3_metadata.shape,
+        dtype=v3_metadata.data_type.to_numpy(),
+        chunks=v3_metadata.chunks,
+        fill_value=fill_value or v3_metadata.fill_value,
+        compressor=compressor_config,
+        filters=filter_configs,
+        order="C",
+        attributes=v3_metadata.attributes,
+        dimension_separator=".",  # Assuming '.' as default dimension separator
+    )
+    return v2_metadata
+
+
 def variable_to_kerchunk_arr_refs(var: Variable, var_name: str) -> KerchunkArrRefs:
     """
     Create a dictionary containing kerchunk-style array references from a single xarray.Variable (which wraps either a ManifestArray or a numpy array).
@@ -95,6 +148,7 @@ def variable_to_kerchunk_arr_refs(var: Variable, var_name: str) -> KerchunkArrRe
     Partially encodes the inner dicts to json to match kerchunk behaviour (see https://github.com/fsspec/kerchunk/issues/415).
     """
     from virtualizarr.manifests import ManifestArray
+    from virtualizarr.translators.kerchunk import to_kerchunk_json
 
     if isinstance(var.data, ManifestArray):
         marr = var.data
